Removes the MemoryMapper, Segments
Removed the memory mapper and segments since we're going to a flat
address space (simplifies things), and we'll bring in the GC at a later
time.

This is being done because I discovered a way we can support @ticking's
image system without needing an indirection map:
1. Flat address space
2. Load the image creating new objects
3. Install forward pointers
4. Let the GC resolve the issue on its next collection cycle (we can
force it as the last phase of the image loading).

This will not affect other types of applications, except make them
quicker. :)