Architectural changes: Register Machine.
Caribou has been derived around a stack based model since inception.
This is changing with this commit. From this point forward, we are
adopting a register based model. We will not be doing away with the
stack, but we'll be adding registers.

This is being done so that we can avoid unnecessary pushing/popping
of operands in the case of nested functions, for instance. It will
complicate the compiler a little, but the performance characteristics
are worth an extra bit of complexity.